+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_manager.py
@@ -124,7 +124,7 @@ class BinanceWebSocketApiManager(threading.Thread):
                  throw_exception_if_unrepairable=False,
                  restart_timeout=6):
         threading.Thread.__init__(self)
-        self.version = "1.16.6.dev"
+        self.version = "1.16.7"
         logging.info("New instance of unicorn_binance_websocket_api_manager " + self.version + " started ...")
         colorama.init()
         if process_stream_data is False: